Stabilize libc type and macro ABI boundaries

+++ include/unistd.h
@@ -1,9 +1,12 @@
 #ifndef _UNISTD_H
 #define _UNISTD_H
 
-long read(int fd, void *buf, unsigned long count);
-long write(int fd, const void *buf, unsigned long count);
-long lseek(int fd, long offset, int whence);
+#include <stddef.h>
+#include <sys/types.h>
+
+ssize_t read(int fd, void *buf, size_t count);
+ssize_t write(int fd, const void *buf, size_t count);
+off_t lseek(int fd, off_t offset, int whence);
 int close(int fd);
 
 #endif
